@@ -1,10 +1,16 @@
 {
   "name": "nor-extend",
-  "version": "1.1.1",
-  "description": "Extends Q promises with custom methods from another object",
-  "main": "src/index.js",
+  "version": "1.3.0",
+  "description": "Extends promises with custom methods from another object",
+  "main": "dist/index.js",
   "scripts": {
-    "test": "mocha -R spec tests/test-*.js",
+    "compile-watch": "babel --watch -d dist/ src/",
+    "compile": "babel -d dist/ src/",
+    "prepare": "npm run -s compile",
+    "pretest": "npm run -s compile",
+    "test": "mocha --require @babel/register test/unit/**/*.spec.js",
+    "preversion": "npm test",
+    "postversion": "git push && git push --tags",
     "jshint": "jshint src/"
   },
   "directories": {
@@ -27,18 +33,20 @@
     "url": "https://github.com/sendanor/nor-extend/issues"
   },
   "dependencies": {
-    "q": "^1.0.1",
-    "nor-is": "^1.0.1",
-    "nor-debug": "^1.0.1",
-    "nor-array": "^1.0.1",
-    "nor-function": "^1.0.1"
+    "@norjs/is": "^1.3.0",
+    "@norjs/debug": "^1.3.0"
   },
   "devDependencies": {
-    "jshint": "^2.9.2",
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/register": "^7.0.0",
+    "jshint": "^2.9.7",
     "lint": "^1.1.2",
-    "mocha": "^5.2.0"
+    "mocha": "^5.2.0",
+    "q": "^1.5.1"
   },
   "engines": {
-    "node": ">=0.8"
+    "node": ">=4"
   }
 }
