@@ -1,3 +1,5 @@
 "use strict";
-var mod = module.exports = require('./extend.js');
-mod.ActionObject = require('./ActionObject.js');
+export * from'./extend.js';
+
+import Extend from './extend.js';
+export default Extend;
