@@ -1,8 +1,8 @@
 "use strict";
 
-var extend = require('nor-extend');
-var assert = require('assert');
-var Q = require('q');
+import extend from 'nor-extend';
+import assert from 'assert';
+import Q from 'q';
 
 function get_names(obj) {
 	return Object.getOwnPropertyNames(obj);
@@ -50,9 +50,9 @@ describe('extend', function(){
 			}
 			Foo.prototype.foo = function() { return 1; }
 			Foo.prototype.bar = function() { return 2; }
-			var foo = new Foo();
-			var obj = ["hello", "world"];
-			var extended_obj = extend.object(foo, extend.getMethodNamesFromObject(foo), obj);
+			let foo = new Foo();
+			let obj = ["hello", "world"];
+			let extended_obj = extend.object(foo, extend.getMethodNamesFromObject(foo), obj);
 			assert.strictEqual( extended_obj[0], "hello" );
 			assert.strictEqual( extended_obj[1], "world" );
 			assert.strictEqual( extended_obj.foo(), 1 );
@@ -64,9 +64,9 @@ describe('extend', function(){
 			}
 			Foo.prototype.foo = function() { return 1; }
 			Foo.prototype.bar = function() { return 2; }
-			var foo = new Foo();
-			var obj = ["hello", "world"];
-			var extended_obj = extend.object(foo, Foo, obj);
+			let foo = new Foo();
+			let obj = ["hello", "world"];
+			let extended_obj = extend.object(foo, Foo, obj);
 			assert.strictEqual( extended_obj[0], "hello" );
 			assert.strictEqual( extended_obj[1], "world" );
 			assert.strictEqual( extended_obj.foo(), 1 );
@@ -77,9 +77,9 @@ describe('extend', function(){
 			function Foo() { }
 			Foo.prototype.foo = function() { return 1; }
 			Foo.prototype.bar = function() { return 2; }
-			var foo = new Foo();
-			var obj = ["hello", "world"];
-			var extended_obj = extend.object(foo, obj);
+			let foo = new Foo();
+			let obj = ["hello", "world"];
+			let extended_obj = extend.object(foo, obj);
 			assert.strictEqual( extended_obj[0], "hello" );
 			assert.strictEqual( extended_obj[1], "world" );
 			assert.strictEqual( extended_obj.foo(), 1 );
@@ -94,9 +94,9 @@ describe('extend', function(){
 			function Foo() {
 			}
 			Foo.prototype.foo = function() { return 1; }
-			var foo = new Foo();
-			var obj = ["hello", "world"];
-			var extended_obj = extend.object(foo, [Foo, Bar], obj);
+			let foo = new Foo();
+			let obj = ["hello", "world"];
+			let extended_obj = extend.object(foo, [Foo, Bar], obj);
 			assert.strictEqual( extended_obj[0], "hello" );
 			assert.strictEqual( extended_obj[1], "world" );
 			assert.strictEqual( extended_obj.foo(), 1 );
@@ -114,14 +114,14 @@ describe('extend', function(){
 			Foobar.prototype.foo = function(x) { return x+1; }
 			Foobar.prototype.bar = function(x) { return x+2; }
 
-			var defer = Q.defer();
+			let defer = Q.defer();
 			setTimeout(function() {
 				defer.resolve( new Foobar() );
 			}, 200);
-			var p = defer.promise;
+			let p = defer.promise;
 
-			var methods = extend.getMethodNamesFromConstructor(Foobar);
-			var extended_p = extend.promise( methods, p);
+			let methods = extend.getMethodNamesFromConstructor(Foobar);
+			let extended_p = extend.promise( methods, p);
 
 			assert.strictEqual( typeof extended_p, 'object' );
 			assert.strictEqual( typeof extended_p.foo, 'function' );
@@ -137,14 +137,14 @@ describe('extend', function(){
 
 		it('should extend Q promise with methods from Array', function(done){
 
-			var defer = Q.defer();
+			let defer = Q.defer();
 			setTimeout(function() {
 				defer.resolve( ["hello", "world", "foo", "bar"] );
 			}, 200);
-			var p = defer.promise;
+			let p = defer.promise;
 
-			var methods = extend.getMethodNamesFromConstructor(Array);
-			var extended_p = extend.promise( methods, p);
+			let methods = extend.getMethodNamesFromConstructor(Array);
+			let extended_p = extend.promise( methods, p);
 
 			assert.strictEqual( typeof extended_p, 'object' );
 			assert.strictEqual( typeof extended_p.shift, 'function' );
@@ -188,14 +188,14 @@ describe('extend', function(){
 				return this.x;
 			};
 
-			var defer = Q.defer();
+			let defer = Q.defer();
 			setTimeout(function() {
 				defer.resolve( new Foobar(1000) );
 			}, 200);
-			var p = defer.promise;
+			let p = defer.promise;
 
-			var methods = extend.getMethodNamesFromConstructor(Foobar);
-			var extended_p = extend.promise( methods, p);
+			let methods = extend.getMethodNamesFromConstructor(Foobar);
+			let extended_p = extend.promise( methods, p);
 
 			assert.strictEqual( typeof extended_p, 'object' );
 			assert.strictEqual( typeof extended_p.inc, 'function' );
@@ -222,7 +222,7 @@ describe('extend', function(){
 			}
 
 			Foobar.create = function(x) {
-				var defer = Q.defer();
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						defer.resolve( new Foobar(x) );
@@ -234,8 +234,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.inc = function(x) {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof self.x, 'number');
@@ -251,8 +251,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.dec = function(x) {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof self.x, 'number');
@@ -268,8 +268,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.get = function() {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof self.x, 'number');
@@ -282,10 +282,10 @@ describe('extend', function(){
 				return defer.promise;
 			};
 
-			var p = Foobar.create(1000);
+			let p = Foobar.create(1000);
 
-			var methods = extend.getMethodNamesFromConstructor(Foobar);
-			var extended_p = extend.promise( methods, p);
+			let methods = extend.getMethodNamesFromConstructor(Foobar);
+			let extended_p = extend.promise( methods, p);
 
 			assert.strictEqual( typeof extended_p, 'object' );
 			assert.strictEqual( typeof extended_p.inc, 'function' );
@@ -310,7 +310,7 @@ describe('extend', function(){
 			}
 
 			Foobar.create = function(x) {
-				var defer = Q.defer();
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						defer.resolve( new Foobar(x) );
@@ -322,8 +322,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.inc = function(x) {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof self.x, 'number');
@@ -339,8 +339,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.dec = function(x) {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof self.x, 'number');
@@ -356,8 +356,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.get = function() {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof self.x, 'number');
@@ -370,7 +370,7 @@ describe('extend', function(){
 				return extend.promise( extend.getMethodNamesFromConstructor(Foobar), defer.promise);
 			};
 
-			var p = Foobar.create(1000);
+			let p = Foobar.create(1000);
 
 			assert.strictEqual( typeof p, 'object' );
 			assert.strictEqual( typeof p.inc, 'function' );
@@ -397,7 +397,7 @@ describe('extend', function(){
 			}
 
 			Foobar.create = function(x) {
-				var defer = Q.defer();
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						defer.resolve( new Foobar(x) );
@@ -409,8 +409,8 @@ describe('extend', function(){
 			};
 
 			Foobar.prototype.query = function(x) {
-				var self = this;
-				var defer = Q.defer();
+				let self = this;
+				let defer = Q.defer();
 				setTimeout(function() {
 					try {
 						assert.strictEqual(typeof x, 'number');
@@ -424,7 +424,7 @@ describe('extend', function(){
 				return extend.promise( Array, defer.promise);
 			};
 
-			var p = Foobar.create(1000);
+			let p = Foobar.create(1000);
 
 			assert.strictEqual( typeof p, 'object' );
 			assert.strictEqual( typeof p.$query, 'function' );
@@ -444,16 +444,16 @@ describe('extend', function(){
 	// extend.copy(obj)
 	describe('.copy()', function(){
 		it('should copy an object', function(){
-			var obj = {'foo':'bar'};
-			var obj2 = extend.copy(obj);
+			let obj = {'foo':'bar'};
+			let obj2 = extend.copy(obj);
 			assert.strictEqual( obj2.foo, obj.foo );
 			obj.foo = 'hello world';
 			assert.strictEqual( obj.foo, 'hello world' );
 			assert.strictEqual( obj2.foo, 'bar' );
 		});
 		it('should deep copy an object', function(){
-			var obj = {'foo':{'hello':'world'}};
-			var obj2 = extend.copy(obj);
+			let obj = {'foo':{'hello':'world'}};
+			let obj2 = extend.copy(obj);
 			assert.strictEqual( obj2.foo.hello, 'world' );
 			obj.foo.hello = 'hello world';
 			assert.strictEqual( obj.foo.hello, 'hello world' );
